requirements.txt
annotated-types==0.7.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gspread==6.2.1
httplib2==0.31.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
oauthlib==3.3.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
shellingham==1.5.4
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
fastapi>=0.109.0
uvicorn>=0.27.0
